Add writeUTF test and reorganize writing tests
