+++ src/test/java/me/hugmanrique/jacobin/writer/BigEndianByteStreamWriterTest.java
@@ -1,0 +1,14 @@
+package me.hugmanrique.jacobin.writer;
+
+import java.nio.ByteOrder;
+
+/**
+ * @author Hugo Manrique
+ * @since 22/12/2018
+ */
+public class BigEndianByteStreamWriterTest extends EndiannessByteStreamWriterTest {
+
+    public BigEndianByteStreamWriterTest() {
+        super(ByteOrder.BIG_ENDIAN);
+    }
+}
